RTreeMapImporter + Code cleanup
Removed RTreeMap for now as well
