+++ inc/LinkDef.h
@@ -2,6 +2,5 @@
 #pragma link off all globals;
 #pragma link off all classes;
 #pragma link off all functions;
-#pragma link C++ class TreeMap+;
 #pragma link C++ class TTreeMap+;
 #endif
