+++ inc/TTreeMap.hxx
@@ -4,7 +4,7 @@
 #ifndef TTREEMAP_HXX
 #define TTREEMAP_HXX
 
-#include "TreeMap.hxx"
+#include "RTreeMapBase.hxx"
 
 #include "TROOT.h"
 #include "TObject.h"
@@ -13,20 +13,18 @@
 
 #include <vector>
 
-class TTreeMap final : public TreeMap, public TObject {
+class TTreeMap final : public ROOT::Experimental::RTreeMapBase, public TObject {
 public:
-   struct Node final : public TreeMap::Node, public TObject {
+   struct Node final : public ROOT::Experimental::RTreeMapBase::Node, public TObject {
    public:
       ClassDef(Node, 1)
    };
-
-   TTreeMap(const ROOT::Experimental::RNTupleInspector &insp) : TreeMap(insp) {}
    TTreeMap() = default;
    void Paint(Option_t *opt) override;
 
-   ClassDefOverride(TTreeMap, 1)
-private:
-   void AddBox(const Rect &rect, const RGBColor &color, float borderWidth) const final;
-   void AddText(const Vec2 &pos, const std::string &content, float size, const RGBColor &color=RGBColor(0,0,0), bool alignCenter=false) const final;
+   ClassDefOverride(TTreeMap, 1) private
+      : void AddBox(const Rect &rect, const RGBColor &color, float borderWidth) const final;
+   void AddText(const Vec2 &pos, const std::string &content, float size, const RGBColor &color = RGBColor(0, 0, 0),
+                bool alignCenter = false) const final;
 };
-#endif // TTREEMAP_HXX
+#endif
