+++ main.cpp
@@ -1,5 +1,5 @@
 #include "TTreeMap.hxx"
-#include "RTreeMap.hxx"
+#include "RTreeMapImporter.hxx"
 
 #include <ROOT/RCanvas.hxx>
 #include <ROOT/RBrowser.hxx>
@@ -13,30 +13,21 @@
 
 static const std::string kTupleName = "poster_tuple";
 static const std::string kFileName = "../example_file.root";
-void DisplayRCanvas()
-{
-   const auto insp = ROOT::Experimental::RNTupleInspector::Create(kTupleName, kFileName);
-   auto canvas = ROOT::Experimental::RCanvas::Create("RNTupleBrowser");
-   canvas->Draw<ROOT::Experimental::RTreeMap>(canvas, std::cref(*insp));
-   canvas->Show();
-}
-void DisplayTCanvas(bool disableJs=false)
+void DisplayCanvas(bool disableJs=false)
 {
   if(!disableJs) {
   ROOT::RWebWindowsManager::AddServerLocation("webcanv", "../webcanv");
   TWebCanvas::AddCustomClass("TTreeMap");
   TWebCanvas::SetCustomScripts("modules:webcanv/treemap.mjs");
   }
-  const auto insp = ROOT::Experimental::RNTupleInspector::Create(kTupleName, kFileName);
-  auto tm = new TTreeMap(*insp);
+  auto importer = ROOT::Experimental::RTreeMapImporter::Create(kFileName, kTupleName);
   auto c = new TCanvas("c_tm","TreeMap");
-  c->Add(tm);
+  c->Add(importer->Import());
 }
 int main(int argc, char **argv)
 {
    TApplication app("ROOT Application", &argc, argv);
-   //DisplayRCanvas();
-   DisplayTCanvas();
+   DisplayCanvas(true);
    app.Run();
    return 0;
 }
